DimensionedField: Templated on the primitive field type
The template arguments have changed from:

    template
    <
        class Type,
        class GeoMesh
    >

to:

    template
    <
        class Type,
        class GeoMesh,
        template<class> class PrimitiveField = Field
    >

The new PrimitiveField argument specifies which primitive field class
(i.e., without geometry or dimensions) is to be used to store the data.
This argument defaults to the standard Field. Other possibilities
include DynamicField and SubField.

All existing code uses the standard Field. So, code that uses the
DimensionedField template directly can rely on the default
PrimitiveField argument, and therefore does not to be changed in any
way.