GeometricField: Templated on the primitive field type
The template arguments have changed from:

    template
    <
        class Type,
        template<class> class PatchField,
        class GeoMesh
    >

to:

    template
    <
        class Type,
        class GeoMesh,
        template<class> class PrimitiveField = Field
    >

The new PrimitiveField argument specifies which primitive field class
(i.e., without geometry or dimensions) is to be used to store the data.
This argument defaults to the standard Field. Other possibilities
include DynamicField and SubField.

The PatchField argument is no longer needed as it is now determined
automatically from the type of the GeoMesh.

All existing code uses the standard Field. So, code that uses the
GeometricField template directly can rely on the default PrimitiveField
argument, and threfore does not need to specify the PrimitiveField
explicitly. The only change needed is the removal of the PatchField
argument. Code that uses more specific type definitions, such as
volScalarField, surfaceVectorField::Internal and VolField<Type> will not
need changing.