fvMeshTopoChangers::meshToMesh: Maintain valid meshes during mapping
The current mesh is now swapped with the new mesh prior to the mapping
of fields and other properties. Previously the new mesh was copied into
the current mesh.

This change means that both meshes are valid during the mapping
operation, and properties of either can be used. It should also now be
be more efficient as a swap operation just exchanges list pointers and
sizes, whilst a copy requires duplicating all the primitive mesh data.